Use custom collectors for worker and queue metrics
The best practice for exporters is to collect as many metrics as
possible at the time Prometheus does the scrape. This fixes issues where
threads or connections may die in the background and cause stale metrics
to be shown forever. Instead, the scrape will fail and we can alert if
celery exporter is down.